09 ex01 done

+++ 09_Module/ex01/RPN.hpp
@@ -25,7 +25,7 @@ class RPN
 		void Calculate(std::string str);
 
 	private:
-		std::stack<long> stack;
+		std::stack<float> stack;
 };
 
 #endif
